CaptureProcess class created
