+++ server/CEC.py
@@ -1,138 +1,24 @@
 # IMPORTS >>> DO NOT CHANGE <<<
 import warnings
 warnings.filterwarnings('ignore')
-import socket,time,math,click, os
+import click, os, math
 from datetime import datetime
 import numpy as np
 from scipy.interpolate import CubicSpline
-from cv2 import destroyAllWindows,triangulatePoints
+from cv2 import destroyAllWindows, triangulatePoints
 
+from mcr.CaptureProcess import CaptureProcess
 from mcr.math import isCollinear
 from mcr.cameras import estimateFundMatrix_8norm, decomposeEssentialMat, projectionPoints
 from mcr.markers import orderCenterCoord, occlusion, processCentroids
 from mcr.plot import plotArena
-from mcr.constants import cameraMat, distCoef
 
-class CEC(object):
-    def __init__(self,cameraids, markers,trigger,record,fps,verbose,save):
-        # VARIABLES >>> DO NOT CHANGE <<<
-        self.cameraids = str(cameraids).split(',')
-        self.cameras = len(self.cameraids)
-        self.markers = markers
-        self.trigger = trigger
-        self.record = record
-        self.fps = fps
-        self.step = 1 / fps
-        self.verbose = verbose
-        self.save = save
-        self.ipList = []
-
-        # IP lookup from hostname
-        try:
-            self.ipList = [socket.gethostbyname(f'cam{idx}.local') for idx in self.cameraids]
-        except socket.gaierror as e:
-            print('[ERROR] Number of cameras do not match the number of IPs found')
-            exit()
-
-        self.cameraMat = np.copy(cameraMat)
-        self.distCoef = np.copy(distCoef)
-
-        # Do not change below this line, socket variables
-        self.nImages = int(self.record / self.step)
-        self.imageSize = []
-        
-        for _ in range(self.cameras): 
-            self.imageSize.append([])
-        
-        print('[INFO] creating server')
-
-        self.bufferSize = 1024
-        self.server_socket = socket.socket(family=socket.AF_INET, type=socket.SOCK_DGRAM) 
-        self.server_socket.bind(('0.0.0.0',8888))
-
-    # Connect with clients
-    def connect(self):
-        print('[INFO] server running, waiting for clients')
-
-        addedCams,ports=[],[]
-        while len(addedCams)!=self.cameras:
-            # Collect adresses
-            message,address = self.server_socket.recvfrom(self.bufferSize)
-
-            # Check if it is in IP list
-            if address[0] not in self.ipList:
-                print('[ERROR] IP '+address[0]+' not in the list')
-                exit()
-
-            # Get image size
-            idx = self.ipList.index(address[0])
-            self.imageSize[idx] = np.array(message.decode('utf-8').split(',')).astype(np.int)
-            print('[INFO] camera '+str(idx)+' connected at '+str(address[0]))
-
-            # Redo intrinsics
-            ret,newCamMatrix=self.mcrIntrinsics(self.cameraMat[idx],self.imageSize[idx][0],self.imageSize[idx][1],self.imageSize[idx][2])
-            if ret: 
-                self.cameraMat[idx]=np.copy(newCamMatrix)
-            else: 
-                exit()
-
-            addedCams.append(idx)
-            ports.append(address)
-        
-        print('[INFO] all clients connected')
-
-        # Send trigger
-        self.trigger += time.time()
-        for i in range(self.cameras): 
-            self.server_socket.sendto((str(self.trigger)+' '+str(self.record)).encode(),tuple(ports[i]))
-        print('[INFO] trigger sent')
-
-    # New intrinsics
-    def mcrIntrinsics(self,origMatrix,w,h,mode):
-        camIntris = np.copy(origMatrix) # Copy to avoid register error
-
-        # Check if image is at the available proportion
-        if w/h==4/3 or w/h==16/9:
-            if mode==4: # Only resize
-                ratio = w/960
-                camIntris[0][0],camIntris[0][2]=ratio*camIntris[0][0],ratio*camIntris[0][2]
-                camIntris[1][1],camIntris[1][2]=ratio*camIntris[1][1],ratio*camIntris[1][2]
-
-            elif mode==5: # Crop in X and resize
-                ratio = 1640/960
-                camIntris[0][0],camIntris[0][2]=ratio*camIntris[0][0],ratio*camIntris[0][2]
-                camIntris[1][1],camIntris[1][2]=ratio*camIntris[1][1],ratio*camIntris[1][2]-155
-                ratio = w/1640
-                camIntris[0][0],camIntris[0][2]=ratio*camIntris[0][0],ratio*camIntris[0][2]
-                camIntris[1][1],camIntris[1][2]=ratio*camIntris[1][1],ratio*camIntris[1][2]
-            
-            elif mode==6: # Crop in Y and X and resize
-                ratio=1640/960
-                camIntris[0][0],camIntris[0][2]=ratio*camIntris[0][0],ratio*camIntris[0][2]-180
-                camIntris[1][1],camIntris[1][2]=ratio*camIntris[1][1],ratio*camIntris[1][2]-255
-                ratio = w/1280
-                camIntris[0][0],camIntris[0][2]=ratio*camIntris[0][0],ratio*camIntris[0][2]
-                camIntris[1][1],camIntris[1][2]=ratio*camIntris[1][1],ratio*camIntris[1][2]
-
-            elif mode==7: # Crop in Y and X and resize
-                ratio=1640/960
-                camIntris[0][0],camIntris[0][2]=ratio*camIntris[0][0],ratio*camIntris[0][2]-500
-                camIntris[1][1],camIntris[1][2]=ratio*camIntris[1][1],ratio*camIntris[1][2]-375
-                ratio = w/640
-                camIntris[0][0],camIntris[0][2]=ratio*camIntris[0][0],ratio*camIntris[0][2]
-                camIntris[1][1],camIntris[1][2]=ratio*camIntris[1][1],ratio*camIntris[1][2]
-            else:
-                print('[ERROR] conversion for intrinsics matrix not known')
-                return False,camIntris
-            return True,camIntris
-        else:
-            print('[ERROR] out of proportion of the camera mode')
-            return False,camIntris
-
-    # Collect points from clients, order and trigger interpolation
+class CEC(CaptureProcess):
+    # Collect points from clients, order and trigger interpolation for Camera Extrinsics Calibration
     def collect(self):
-        # Internal variables
         print('[INFO] waiting capture')
+       
+        # Internal variables
         capture = np.ones(self.cameras,dtype=np.bool)
         counter,lastTime = np.zeros(self.cameras,dtype=np.uint16),np.zeros(self.cameras,dtype=np.uint32)
         missed,invalid = np.zeros(self.cameras,dtype=np.uint32),np.zeros(self.cameras,dtype=np.uint32)
@@ -174,7 +60,7 @@ class CEC(object):
                         a,b,time,imgNumber = message[-4],message[-3],message[-2],int(message[-1]) 
 
                         # Undistort points
-                        undCoord = processCentroids(coord,a,b,self.cameraMat[idx],distCoef[idx])
+                        undCoord = processCentroids(coord,a,b,self.cameraMat[idx],self.distCoef[idx])
                         
                         if self.save: dfSave.append(np.concatenate((undCoord.reshape(6),[time,imgNumber,idx])))
 
@@ -230,19 +116,6 @@ class CEC(object):
             self.server_socket.close()
             destroyAllWindows()
 
-            # Save Camera Extrinsics Calibration (CEC) Data
-            if self.save: 
-                now = datetime.now()
-                DMY, HMS = now.strftime('%d-%m-%y'), now.strftime('%H-%M-%S')
-                path = 'dataSaves/' + DMY + '/'
-
-                # Check whether directory already exists
-                if not os.path.exists(path):
-                    os.mkdir(path)
-                    print('Folder %s created!' % path)
-        
-                np.savetxt(path + 'CEC-' + HMS + '.csv', np.array(dfSave), delimiter=',')
-
             # Get last interval
             for idx in range(self.cameras):
                 if not len(dfOrig[idx]): continue
@@ -254,6 +127,19 @@ class CEC(object):
             # Print results
             print('[RESULTS] server results are')
             for i in range(self.cameras): print('  >> camera '+str(i)+': '+str(len(dfOrig[i]))+' valid images, address '+str(self.ipList[i])+', missed '+str(int(missed[i]))+' images')
+            
+            # Save Camera Extrinsics Calibration (CEC) Data
+            if self.save: 
+                now = datetime.now()
+                DMY, HMS = now.strftime('%d-%m-%y'), now.strftime('%H-%M-%S')
+                path = './dataSaves/' + DMY + '/'
+
+                # Check whether directory already exists
+                if not os.path.exists(path):
+                    os.mkdir(path)
+                    print('Folder %s created!' % path)
+        
+                np.savetxt(path + 'CEC-' + HMS + '.csv', np.array(dfSave), delimiter=',')
             
             # Get pose between each pair
             for cam in range(self.cameras-1):
